retail_audit_django/requirements.txt
django>=4.2
ultralytics
torch
torchvision
numpy
pillow
opencv-python-headless
psycopg2-binary    
pgvector